+++ src/components/windows/main_screen.py
@@ -19,7 +19,7 @@ class MainScreen(tk.Frame):
         """Creates the widgets of the main screen.
         """
         title_label = tk.Label(self, text=self.controller.config.app_name, font=("Arial", 15))
-        title_label.pack(pady=20)
+        title_label.pack(pady=6)
 
         separator_frame = tk.Canvas(self, height=1, width=250, bg='dark grey')
         separator_frame.pack(pady=(0, 10))